+++ smart-admin-api/sa-admin/src/main/java/net/lab1024/sa/admin/module/system/TwAdmin/service/impl/TwUserCoinServiceImpl.java
@@ -21,6 +21,7 @@ import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
 
 import javax.annotation.Resource;
+import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -36,11 +37,11 @@ public class TwUserCoinServiceImpl extends ServiceImpl<TwUserCoinDao, TwUserCoin
     @Lazy
     private TwUserService twUserService;
     @Override
-    public int incre(Integer uid, double num,String coinname) {
+    public int incre(Integer uid, BigDecimal num, String coinname) {
         return this.baseMapper.incre(uid,num,coinname);
     }
 
-    public int decre(Integer uid, double num,String coinname) {
+    public int decre(Integer uid, BigDecimal num,String coinname) {
         return this.baseMapper.decre(uid,num,coinname);
     }
